+++ modi/seller-service/src/main/java/com/coc/modi/seller/chat/exception/ChatAccessDeniedException.java
@@ -1,8 +1,8 @@
-package com.coc.modi.seller.exception;
+package com.coc.modi.seller.chat.exception;
 
 import com.coc.modi.common.ErrorCode;
 
-public class ChatAccessDeniedException extends SellerException {
+public class ChatAccessDeniedException extends ChatException {
 
 	public ChatAccessDeniedException(String detailMessage) {
 		super(ErrorCode.FORBIDDEN, detailMessage);
