+++ modi/seller-service/src/main/java/com/coc/modi/seller/chat/exception/ChatInputInvalidException.java
@@ -1,0 +1,10 @@
+package com.coc.modi.seller.chat.exception;
+
+import com.coc.modi.common.ErrorCode;
+
+public class ChatInputInvalidException extends ChatException {
+
+	public ChatInputInvalidException(String detailMessage) {
+		super(ErrorCode.INVALID_INPUT, detailMessage);
+	}
+}
