+++ modi/seller-service/src/main/java/com/coc/modi/seller/chat/exception/ChatRoomNotFoundException.java
@@ -1,8 +1,8 @@
-package com.coc.modi.seller.exception;
+package com.coc.modi.seller.chat.exception;
 
 import com.coc.modi.common.ErrorCode;
 
-public class ChatRoomNotFoundException extends SellerException {
+public class ChatRoomNotFoundException extends ChatException {
 
 	public ChatRoomNotFoundException(String detailMessage) {
 		super(ErrorCode.NOT_FOUND, detailMessage);
