+++ modi/seller-service/src/main/java/com/coc/modi/seller/seller/application/SellerService.java
@@ -1,8 +1,8 @@
 package com.coc.modi.seller.seller.application;
 
 import com.coc.modi.seller.seller.application.dto.SellerRentalResponse;
-import com.coc.modi.seller.exception.SellerDuplicateException;
-import com.coc.modi.seller.exception.SellerNotFoundException;
+import com.coc.modi.seller.seller.exception.SellerDuplicateException;
+import com.coc.modi.seller.seller.exception.SellerNotFoundException;
 import com.coc.modi.seller.seller.infrastructure.client.member.MemberClientAdapter;
 import com.coc.modi.seller.seller.application.dto.SellerCreateCommand;
 import com.coc.modi.seller.seller.application.dto.SellerDetailResponse;
