+++ modi/seller-service/src/main/java/com/coc/modi/seller/settlement/domain/SellerSettlement.java
@@ -1,8 +1,8 @@
 package com.coc.modi.seller.settlement.domain;
 
 import com.coc.modi.common.BaseEntity;
-import com.coc.modi.seller.exception.SellerSettlementStatusConflictException;
-import com.coc.modi.seller.exception.SettlementInputInvalidException;
+import com.coc.modi.seller.settlement.exception.SellerSettlementStatusConflictException;
+import com.coc.modi.seller.settlement.exception.SettlementInputInvalidException;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
